more changes

+++ src/main/java/SparkTesting/DistributedChiSquareComputer.java
@@ -14,10 +14,10 @@ import scala.Tuple2;
 import com.google.common.collect.Lists;
 
 @SuppressWarnings("serial")
-public class DistributedChiSquareComputer implements Serializable, FlatMapFunction<Iterator<Tuple2<Integer,Tuple2<Double,Block>>>, ChisquareUnit> {
+public class DistributedChiSquareComputer implements Serializable, FlatMapFunction<Iterator<Tuple2<Integer,Tuple2<BigDecimal,Block>>>, ChisquareUnit> {
 
-	public Iterable<ChisquareUnit> call(Iterator<Tuple2<Integer, Tuple2<Double, Block>>> t) throws Exception {
-		List<Tuple2<Double, Block>> rawList = Lists.newArrayList();
+	public Iterable<ChisquareUnit> call(Iterator<Tuple2<Integer, Tuple2<BigDecimal, Block>>> t) throws Exception {
+		List<Tuple2<BigDecimal, Block>> rawList = Lists.newArrayList();
 		List<ChisquareUnit> returnList = Lists.newArrayList();
 		
 		while(t.hasNext()) {
@@ -26,10 +26,10 @@ public class DistributedChiSquareComputer implements Serializable, FlatMapFuncti
 		// Lets sort the data again. This is because when the data was added to the previous partition,
 		// the sorted arrangement may have been lost. Since the partition fits into the worker's memory
 		// we don't have to create a RDD to sort the partition data.
-		Collections.sort(rawList, new Comparator<Tuple2<Double, Block>>() {
+		Collections.sort(rawList, new Comparator<Tuple2<BigDecimal, Block>>() {
 
-			public int compare(Tuple2<Double, Block> o1, Tuple2<Double, Block> o2) {
-				return BigDecimal.valueOf(o1._1()).compareTo(BigDecimal.valueOf(o2._1()));
+			public int compare(Tuple2<BigDecimal, Block> o1, Tuple2<BigDecimal, Block> o2) {
+				return o1._1().compareTo(o2._1());
 			}
 		});
